+++ src/vis/bar/barComponents/XAxis.tsx
@@ -12,7 +12,7 @@ import {
   VIS_TICK_LABEL_SIZE_SMALL,
 } from '../../general/constants';
 import { ESortStates, SortIcon } from '../../general/SortIcon';
-import { getLabelOrUnknown } from '../../utils';
+import { getLabelOrUnknown } from '../../general/utils';
 
 function TickText({
   value,
