+++ src/vis/bar/barTypes/StackedBars.tsx
@@ -5,7 +5,7 @@ import ColumnTable from 'arquero/dist/types/table/column-table';
 import { Stack, Text } from '@mantine/core';
 import { SingleBar } from '../barComponents/SingleBar';
 import { EAggregateTypes } from '../../interfaces';
-import { getLabelOrUnknown } from '../../utils';
+import { getLabelOrUnknown } from '../../general/utils';
 
 export function StackedBars({
   groupedTable,
