+++ src/vis/general/LegendItem.tsx
