+++ src/vis/scatter/utils.ts
@@ -21,7 +21,7 @@ import {
 import { getCol } from '../sidebar';
 import { ELabelingOptions, ERegressionLineType, IScatterConfig } from './interfaces';
 import { VIS_LABEL_COLOR, VIS_NEUTRAL_COLOR } from '../general/constants';
-import { getLabelOrUnknown } from '../utils';
+import { getLabelOrUnknown } from '../general/utils';
 
 function truncateString(text: string, maxLength = 20): string {
   return text.length > maxLength ? `${text.substring(0, maxLength - 3)}...` : text;
